requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.0
apache-airflow==2.6.2
apache-airflow-providers-common-sql==1.5.1
apache-airflow-providers-ftp==3.4.1
apache-airflow-providers-http==4.4.1
apache-airflow-providers-imap==3.2.1
apache-airflow-providers-sqlite==3.4.1
apache-airflow-providers-telegram==4.1.0
apispec==5.2.2
apturl==0.5.2
argcomplete==3.1.1
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
Babel==2.12.1
beanie==1.19.2
beautifulsoup4==4.12.2
blinker==1.6.2
bokeh==3.1.1
Brlapi==0.8.3
cachelib==0.9.0
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
cmake==3.26.4
colorama==0.4.6
colorlog==4.8.0
command-not-found==0.3
ConfigUpdater==3.1.1
connexion==2.14.2
constantly==15.1.0
contourpy==1.1.0
cron-descriptor==1.4.0
croniter==1.3.15
cryptography==40.0.2
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.1.1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
dj-rest-auth==4.0.1
Django==4.2.2
django-allauth==0.54.0
django-cors-headers==4.1.0
django-debug-toolbar==4.1.0
django-filter==23.2
django-guardian==2.4.0
django-taggit==4.0.0
djangorestframework==3.14.0
dnspython==2.3.0
docutils==0.20.1
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.97.0
filelock==3.12.2
Flask==2.2.5
Flask-AppBuilder==4.3.1
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.54.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.8.0
hiredis==2.2.3
httpcore==0.16.3
httplib2==0.20.2
httptools==0.5.0
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.6.4
importlib-resources==5.12.0
incremental==22.10.0
inflection==0.5.1
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jax==0.4.12
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jose==1.0.0
jsonschema==4.17.3
keras==2.12.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-model==0.0.5
lazy-object-proxy==1.9.0
libclang==16.0.0
limits==3.5.0
linkify-it-py==2.0.2
lit==16.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.7.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==8.10.0
motor==3.2.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.4
netifaces==0.11.0
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.9.1
packaging==21.3
pandas==2.0.2
parsel==1.8.1
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.8.0
plotly==5.15.0
pluggy==1.0.0
prison==0.2.1
Protego==0.2.1
protobuf==4.23.3
psutil==5.9.5
psycopg2==2.9.6
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.9
PyDispatcher==2.0.7
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.7.0
pymacaroons==0.13.0
pymongo==4.4.0
PyNaCl==1.5.0
pynvim==0.4.2
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.19.3
pyTelegramBotAPI==4.12.0
python-apt==2.4.0+ubuntu1
python-daemon==3.0.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-multipart==0.0.6
python-nvd3==0.15.0
python-slugify==8.0.1
python-telegram-bot==20.3
python3-openid==3.2.0
pytz==2023.3
pytzdata==2020.1
pyxdg==0.27
PyYAML==6.0
queuelib==1.6.2
redis==4.5.5
reportlab==3.6.8
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.4.2
rich_argparse==1.1.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
Scrapy==2.9.0
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==3.3.1
sentry-sdk==1.25.1
service-identity==23.1.0
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
speedtest-cli==2.1.3
SQLAlchemy==1.4.48
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
starlette==0.27.0
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.1.0
tldextract==3.4.4
toml==0.10.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
triton==2.0.0
Twisted==22.10.0
typing_extensions==4.6.3
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.2
ufw==0.36.1
ujson==5.8.0
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.23.1
w3lib==2.1.1
wadllib==1.3.6
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.2.3
whitenoise==6.5.0
wrapt==1.15.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
xyzservices==2023.5.0
yarl==1.9.2
zipp==1.0.0
zope.interface==6.0
asgiref==3.7.2
Django==4.2.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.2
greenlet==2.0.2
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.6
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
SQLAlchemy==2.0.17
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
